+++ R/generate.R
@@ -1,4 +1,29 @@
-#' @title Function \code{evaluate}
+#' @title Function analysis_wildcard
+#' @description Show the analysis wildcard
+#' used in \code{\link{plan_summaries}()}.
+#' @export
+#' @seealso \code{\link{plan_summaries}()}
+#' @return The analysis wildcard used in \code{\link{plan_summaries}()}.
+#' @examples
+#' # See ?plan_analyses for examples
+analysis_wildcard <- function(){
+  "analysis__"
+}
+
+#' @title Function dataset_wildcard
+#' @description Show the dataset wildcard
+#' used in \code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
+#' @export
+#' @seealso \code{\link{plan_analyses}()}
+#' @return The dataset wildcard used in
+#' \code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
+#' @examples
+#' # See ?plan_analyses for examples
+dataset_wildcard <- function(){
+  "dataset__"
+}
+
+#' @title Function \code{evaluate_plan}
 #' @description The commands in workflow plan data frames can have
 #' wildcard symbols that can stand for datasets, parameters, function
 #' arguments, etc. These wildcards can be evaluated over a set of
@@ -44,16 +69,16 @@
 #'   large = simulate(50))
 #' # Create a template workflow plan for the analyses.
 #' methods <- workplan(
-#'   regression1 = reg1(..dataset..),
-#'   regression2 = reg2(..dataset..))
+#'   regression1 = reg1(dataset__),
+#'   regression2 = reg2(dataset__))
 #' # Evaluate the wildcards in the template
 #' # to produce the actual part of the workflow plan
 #' # that encodes the analyses of the datasets.
 #' # Create one analysis for each combination of dataset and method.
-#' evaluate(methods, wildcard = "..dataset..",
+#' evaluate_plan(methods, wildcard = "dataset__",
 #'   values = datasets$target)
 #' # Only choose some combinations of dataset and analysis method.
-#' ans <- evaluate(methods, wildcard = "..dataset..",
+#' ans <- evaluate_plan(methods, wildcard = "dataset__",
 #'   values = datasets$target, expand = FALSE)
 #' ans
 #' # For the complete workflow plan, row bind the pieces together.
@@ -63,14 +88,14 @@
 #' # Each combination of wildcard values will be used
 #' # Except when expand is FALSE.
 #' x <- workplan(draws = rnorm(mean = Mean, sd = Sd))
-#' evaluate(x, rules = list(Mean = 1:3, Sd = c(1, 10)))
-evaluate <- function(
+#' evaluate_plan(x, rules = list(Mean = 1:3, Sd = c(1, 10)))
+evaluate_plan <- function(
   plan,
   rules = NULL,
   wildcard = NULL,
   values = NULL,
   expand = TRUE
-  ){
+){
   if (!is.null(rules)){
     return(evaluations(plan = plan, rules = rules, expand = expand))
   }
@@ -87,7 +112,7 @@ evaluate <- function(
   plan[[major]] <- plan[[minor]]
   matching <- plan[matches, ]
   if (expand){
-    matching <- expand(matching, values)
+    matching <- expand_plan(matching, values)
   }
   values <- rep(values, length.out = nrow(matching))
   matching$command <- Vectorize(
@@ -116,7 +141,7 @@ evaluations <- function(
   }
   stopifnot(is.list(rules))
   for (index in seq_len(length(rules))){
-    plan <- evaluate(
+    plan <- evaluate_plan(
       plan,
       wildcard = names(rules)[index],
       values = rules[[index]],
@@ -126,7 +151,7 @@ evaluations <- function(
   return(plan)
 }
 
-#' @title Function \code{expand}
+#' @title Function \code{expand_plan}
 #' @description Expands a workflow plan data frame by duplicating rows.
 #' This generates multiple replicates of targets with the same commands.
 #' @export
@@ -141,8 +166,8 @@ evaluations <- function(
 #'   large = simulate(50))
 #' # Create replicates. If you want repeat targets,
 #' # this is convenient.
-#' expand(datasets, values = c("rep1", "rep2", "rep3"))
-expand <- function(plan, values = NULL){
+#' expand_plan(datasets, values = c("rep1", "rep2", "rep3"))
+expand_plan <- function(plan, values = NULL){
   if (!length(values)){
     return(plan)
   }
@@ -155,7 +180,7 @@ expand <- function(plan, values = NULL){
   return(plan)
 }
 
-#' @title Function \code{gather}
+#' @title Function \code{gather_plan}
 #' @description Create a new workflow plan data frame with a single new
 #' target. This new target is a list, vector, or other aggregate of
 #' a collection of existing targets in another workflow plan data frame.
@@ -173,19 +198,21 @@ expand <- function(plan, values = NULL){
 #'   small = simulate(5),
 #'   large = simulate(50))
 #' # Create a new target that brings the datasets together.
-#' gather(datasets, target = "my_datasets")
+#' gather_plan(datasets, target = "my_datasets")
 #' # This time, the new target just appends the rows of 'small' and 'large'
 #' # into a single matrix or data frame.
-#' gathered <- gather(datasets, target = "aggregated_data", gather = "rbind")
+#' gathered <- gather_plan(
+#'   datasets, target = "aggregated_data", gather = "rbind"
+#' )
 #' gathered
 #' # For the complete workflow plan, row bind the pieces together.
 #' my_plan <- rbind(datasets, gathered)
 #' my_plan
-gather <- function(
+gather_plan <- function(
   plan = NULL,
   target = "target",
   gather = "list"
-  ){
+){
   command <- paste(plan$target, "=", plan$target)
   command <- paste(command, collapse = ", ")
   command <- paste0(gather, "(", command, ")")
@@ -194,17 +221,17 @@ gather <- function(
   )
 }
 
-#' @title Function \code{analyses}
+#' @title Function \code{plan_analyses}
 #' @description Generate a workflow plan data frame to
 #' analyze multiple datasets using multiple methods of analysis.
-#' @seealso \code{\link{summaries}},
+#' @seealso \code{\link{plan_summaries}},
 #'  \code{\link{make}}, \code{\link{workplan}}
 #' @export
 #' @return An evaluated workflow plan data frame of analysis targets.
 #' @param plan workflow plan data frame of analysis methods.
 #' The commands in the \code{command} column must
-#' have the \code{..dataset..} wildcard where the datasets go.
-#' For example, one command could be \code{lm(..dataset..)}. Then,
+#' have the \code{dataset__} wildcard where the datasets go.
+#' For example, one command could be \code{lm(dataset__)}. Then,
 #' the commands in the output will include \code{lm(your_dataset_1)},
 #' \code{lm(your_dataset_2)}, etc.
 #' @param datasets workflow plan data frame with instructions
@@ -216,34 +243,40 @@ gather <- function(
 #'   large = simulate(50))
 #' # Create a template for the analysis methods.
 #' methods <- workplan(
-#'   regression1 = reg1(..dataset..),
-#'   regression2 = reg2(..dataset..))
+#'   regression1 = reg1(dataset__),
+#'   regression2 = reg2(dataset__))
 #' # Evaluate the wildcards to create the part of the workflow plan
 #' # encoding the analyses of the datasets.
-#' ans <- analyses(methods, datasets = datasets)
+#' ans <- plan_analyses(methods, datasets = datasets)
 #' ans
 #' # For the final workflow plan, row bind the pieces together.
 #' my_plan <- rbind(datasets, ans)
 #' my_plan
-analyses <- function(plan, datasets){
-  evaluate(
+plan_analyses <- function(plan, datasets){
+  plan <- deprecate_wildcard(
+    plan = plan,
+    old = "..dataset..",
+    replacement = dataset_wildcard()
+  )
+  evaluate_plan(
     plan,
-    wildcard = "..dataset..",
+    wildcard = dataset_wildcard(),
     values = datasets$target
   )
 }
 
-#' @title Function \code{summaries}
+#' @title Function \code{plan_summaries}
 #' @description Generate a workflow plan data frame for summarizing
 #' multiple analyses of multiple datasets multiple ways.
-#' @seealso \code{\link{analyses}}, \code{\link{make}}, \code{\link{workplan}}
+#' @seealso \code{\link{plan_analyses}}, \code{\link{make}},
+#' \code{\link{workplan}}
 #' @export
 #' @return An evaluated workflow plan data frame of instructions
 #' for computing summaries of analyses and datasets.
 #' analyses of multiple datasets in multiple ways.
 #' @param plan workflow plan data frame with commands for the summaries.
-#' Use the \code{..analysis..} and \code{..dataset..} wildcards
-#' just like the \code{..dataset..} wildcard in \code{\link{analyses}()}.
+#' Use the \code{analysis__} and \code{dataset__} wildcards
+#' just like the \code{dataset__} wildcard in \code{\link{analyses}()}.
 #' @param analyses workflow plan data frame of analysis instructions
 #' @param datasets workflow plan data frame with instructions to make
 #' or import the datasets.
@@ -258,47 +291,61 @@ analyses <- function(plan, datasets){
 #'   large = simulate(50))
 #' # Create a template workflow plan containing the analysis methods.
 #' methods <- workplan(
-#'   regression1 = reg1(..dataset..),
-#'   regression2 = reg2(..dataset..))
+#'   regression1 = reg1(dataset__),
+#'   regression2 = reg2(dataset__))
 #' # Generate the part of the workflow plan to analyze the datasets.
-#' analyses <- analyses(methods, datasets = datasets)
+#' analyses <- plan_analyses(methods, datasets = datasets)
 #' # Create a template workflow plan dataset with the
 #' # types of summaries you want.
 #' summary_types <- workplan(
-#'   summ = summary(..analysis..),
-#'   coef = coefficients(..analysis..))
+#'   summ = summary(analysis__),
+#'   coef = coefficients(analysis__))
 #' # Evaluate the appropriate wildcards to encode the summary targets.
-#' summaries(summary_types, analyses, datasets, gather = NULL)
-#' summaries(summary_types, analyses, datasets)
-#' summaries(summary_types, analyses, datasets, gather = "list")
-#' summs <- summaries(
+#' plan_summaries(summary_types, analyses, datasets, gather = NULL)
+#' plan_summaries(summary_types, analyses, datasets)
+#' plan_summaries(summary_types, analyses, datasets, gather = "list")
+#' summs <- plan_summaries(
 #'   summary_types, analyses, datasets, gather = c("list", "rbind"))
 #' # For the final workflow plan, row bind the pieces together.
 #' my_plan <- rbind(datasets, analyses, summs)
 #' my_plan
-summaries <- function(
+plan_summaries <- function(
   plan,
   analyses,
   datasets,
   gather = rep("list", nrow(plan))
-  ){
+){
+  plan <- deprecate_wildcard(
+    plan = plan,
+    old = "..analysis..",
+    replacement = analysis_wildcard()
+  )
+  plan <- deprecate_wildcard(
+    plan = plan,
+    old = "..dataset..",
+    replacement = dataset_wildcard()
+  )
   plan <- with_analyses_only(plan)
   out <- plan
   group <- paste(colnames(out), collapse = "_")
   out[[group]] <- out$target
-  if (!any(grepl("..analysis..", out$command, fixed = TRUE))){
+  if (!any(grepl(analysis_wildcard(), out$command, fixed = TRUE))){
     stop(
-      "no '..analysis..' wildcard found in plan$command. ",
-      "Use analyses() instead."
+      "no 'analysis__' wildcard found in plan$command. ",
+      "Use plan_analyses() instead."
       )
   }
-  out <- evaluate(out, wildcard = "..analysis..", values = analyses$target)
-  out <- evaluate(
+  out <- evaluate_plan(
     out,
-    wildcard = "..dataset..",
+    wildcard = analysis_wildcard(),
+    values = analyses$target
+  )
+  out <- evaluate_plan(
+    out,
+    wildcard = dataset_wildcard(),
     values = datasets$target,
     expand = FALSE
-    )
+  )
   if (!length(gather)){
     return(out[setdiff(names(out), group)])
   }
@@ -313,7 +360,7 @@ summaries <- function(
     group,
     function(summary_group){
       summary_type <- summary_group[[group]][1]
-      gather(
+      gather_plan(
         summary_group,
         target = summary_type,
         gather = gather[which(summary_type == plan$target)])
@@ -325,13 +372,13 @@ summaries <- function(
 }
 
 with_analyses_only <- function(plan){
-  has_analysis <- grepl("..analysis..", plan$command, fixed = TRUE)
+  has_analysis <- grepl(analysis_wildcard(), plan$command, fixed = TRUE)
   if (any(!has_analysis)){
     warning(
       "removing ",
       sum(has_analysis),
-      " rows with no ..analysis.. wildcard in the command.",
-      "Use analyses() for these.",
+      " rows with no 'analysis__' wildcard in the command.",
+      "Use plan_analyses() for these.",
       call. = FALSE
     )
   }