+++ R/mclapply.R
@@ -6,7 +6,8 @@ run_mclapply <- function(config){
 worker_mclapply <- function(targets, meta_list, config){
   prune_envir(targets = targets, config = config)
   jobs <- safe_jobs(config$jobs)
-  values <- mclapply(targets, build, meta_list = meta_list,
+  values <- mclapply(
+    X = targets, FUN = drake_build, meta_list = meta_list,
     config = config, mc.cores = jobs)
   assign_to_envir(target = targets, value = values, config = config)
 }