+++ tests/testthat/test-edge-cases.R
@@ -34,10 +34,10 @@ test_with_dir("target conflicts with current import or another target", {
   config <- dbug()
   config$plan <- rbind(config$plan, data.frame(target = "f",
     command = "1+1"))
-  expect_message(check(plan = config$plan,
+  expect_message(check_plan(plan = config$plan,
     envir = config$envir))
   config$plan$target <- "repeated"
-  expect_error(check(plan = config$plan))
+  expect_error(check_plan(plan = config$plan))
 })
 
 test_with_dir("target conflicts with previous import", {
@@ -75,35 +75,4 @@ test_with_dir("true targets can be functions", {
   )
   myfunction <- readd(myfunction)
   expect_equal(myfunction(4), 5)
-})
-
-test_with_dir("deprecation", {
-  expect_warning(default_system2_args(jobs = 1, verbose = FALSE))
-  plan <- data.frame(code = 1:2, output = c("x", "y"))
-  expect_warning(make(plan, verbose = FALSE))
-  expect_warning(make(plan, verbose = FALSE))
-  expect_warning(status())
-  expect_true(is.numeric(readd(x, search = FALSE)))
-  expect_warning(prune(plan[1, ]))
-  expect_equal(cached(), "x")
-  expect_warning(make(workplan(x = 1), return_config = TRUE,
-                      verbose = FALSE))
-  expect_warning(make(workplan(x = 1), clear_progress = TRUE,
-                      verbose = FALSE))
-  expect_silent(make(workplan(x = 1), imports_only = TRUE,
-                     verbose = FALSE))
-  pl1 <- expect_warning(drake::plan(x = 1, y = x))
-  pl2 <- workplan(x = 1, y = x)
-  expect_warning(drake::plan())
-  expect_warning(drake::plan(x = y, file_targets = TRUE))
-  expect_warning(drake::workflow())
-  expect_warning(drake::workflow(x = y, file_targets = TRUE))
-
-  # We need to be able to set the drake version
-  # to check back compatibility.
-  x <- this_cache()
-  x$del(key = "initial_drake_version", namespace = "session")
-  expect_false("initial_drake_version" %in% x$list(namespace = "session"))
-  set_initial_drake_version(cache = x)
-  expect_true("initial_drake_version" %in% x$list(namespace = "session"))
 })